refactor: UserService, AuthService 의존성 정립
