+++ src/main/java/underdogs/devbie/auth/controller/AuthController.java
@@ -2,15 +2,14 @@ package underdogs.devbie.auth.controller;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
+import underdogs.devbie.auth.dto.JwtTokenResponse;
 import underdogs.devbie.auth.service.AuthService;
-import underdogs.devbie.auth.service.dto.UserInfoResponse;
-import underdogs.devbie.user.domain.User;
-import underdogs.devbie.user.service.UserService;
 
 @RestController
 @RequestMapping("/api/oauth")
@@ -18,28 +17,16 @@ import underdogs.devbie.user.service.UserService;
 public class AuthController {
 
     private final AuthService authService;
-    private final UserService userService;
 
     @GetMapping("/login-url")
     public ResponseEntity<String> fetchLoginUrl() {
         return ResponseEntity.ok(authService.fetchLoginUrl());
     }
 
-    @GetMapping("/login")
-    public ResponseEntity<String> login(@RequestParam(value = "code") String code) {
-        // code => access token (github 제공)
-        String accessToken = authService.fetchAccessToken(code);
+    @PostMapping("/login")
+    public ResponseEntity<JwtTokenResponse> login(@RequestParam(value = "code") String code) {
+        JwtTokenResponse jwtTokenResponse = authService.createToken(code);
 
-        // access token => user info
-        UserInfoResponse userInfoResponse = authService.fetchUserInfo(accessToken);
-
-        // user info => user 저장
-        User user = userService.saveOrUpdateUser(userInfoResponse);
-
-        // user info로 토큰 생성
-        String token = authService.createToken(user);
-
-        // 토큰 리턴
-        return ResponseEntity.ok(token);
+        return ResponseEntity.ok(jwtTokenResponse);
     }
 }
