+++ src/main/java/underdogs/devbie/auth/dto/AccessTokenRequest.java
@@ -1,4 +1,4 @@
-package underdogs.devbie.auth.service.dto;
+package underdogs.devbie.auth.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
