+++ src/main/java/underdogs/devbie/auth/service/AuthService.java
@@ -3,11 +3,13 @@ package underdogs.devbie.auth.service;
 import org.springframework.stereotype.Service;
 
 import lombok.RequiredArgsConstructor;
+import underdogs.devbie.auth.dto.JwtTokenResponse;
+import underdogs.devbie.auth.dto.UserInfoResponse;
+import underdogs.devbie.auth.dto.UserTokenDto;
 import underdogs.devbie.auth.jwt.JwtTokenProvider;
 import underdogs.devbie.auth.oauth.GithubClient;
-import underdogs.devbie.auth.service.dto.UserInfoResponse;
-import underdogs.devbie.auth.service.dto.UserTokenDto;
 import underdogs.devbie.user.domain.User;
+import underdogs.devbie.user.service.UserService;
 
 @Service
 @RequiredArgsConstructor
@@ -15,20 +17,20 @@ public class AuthService {
 
     private final GithubClient githubClient;
     private final JwtTokenProvider jwtTokenProvider;
+    private final UserService userService;
 
     public String fetchLoginUrl() {
         return githubClient.fetchLoginUrl();
     }
 
-    public String fetchAccessToken(String code) {
-        return githubClient.fetchAccessToken(code);
-    }
+    public JwtTokenResponse createToken(String code) {
+        String accessToken = githubClient.fetchAccessToken(code);
 
-    public UserInfoResponse fetchUserInfo(String accessToken) {
-        return githubClient.fetchUserInfo(accessToken);
-    }
+        UserInfoResponse userInfoResponse = githubClient.fetchUserInfo(accessToken);
 
-    public String createToken(User user) {
-        return jwtTokenProvider.createToken(UserTokenDto.from(user));
+        User user = userService.saveOrUpdateUser(userInfoResponse);
+
+        String token = jwtTokenProvider.createToken(UserTokenDto.from(user));
+        return JwtTokenResponse.from(token);
     }
 }
