--- src/main/java/underdogs/devbie/controller/HelloController.java
@@ -1,15 +1,0 @@
-package underdogs.devbie.controller;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/api")
-public class HelloController {
-
-    @GetMapping("/hello")
-    public String hello() {
-        return "hello underdogs";
-    }
-}
